+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/CryoAspect.java
@@ -40,8 +40,10 @@ public class CryoAspect extends Enchantment {
 
     public void doPostAttack(@NotNull LivingEntity attacker, @NotNull Entity target, int lvl) {
         int ticks = this.frozenTicksPerLevel(target) + (60 * lvl);
-        if (attacker instanceof Player player)
-            ticks = (int) (ticks * player.getAttackStrengthScale(0.5f));
+        if (attacker instanceof Player player) {
+            float f = player.getAttackStrengthScale(0.5f);
+            ticks = (int) (ticks * f * f);
+        }
         if (target.getTicksFrozen() < ticks)
             target.setTicksFrozen(ticks);
     }